+++ ChangeLog.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Document `-x`/`--xml` and `-t`/`--tap`/`--tap-file` flags in README command line options
+- Document `-x`/`--xml` and `-t`/`--tap` flags in README command line options
 
 ## [4.18.1] - 2026-05-05